+++ config/environments/production.rb
@@ -87,6 +87,8 @@ TaxonWorks::Application.configure do
   # Use default logging formatter so that PID and timestamp are not suppressed.
   config.log_formatter = ::Logger::Formatter.new
 
+  config.action_mailer.default_url_options = { protocol: 'https' }
+
   # Load local settings
   Settings.load_from_settings_file(config, :production)